requirements.txt
ruff
black
pytest
Pillow>=10.0.0
GitPython>=3.1.0
